merge new php changes

+++ ext/mbstring/libmbfl/mbfl/mbfl_memory_device.h
@@ -38,29 +38,33 @@
 
 typedef struct _mbfl_memory_device {
 	unsigned char *buffer;
-	int length;
-	int pos;
-	int allocsz;
+	size_t length;
+	size_t pos;
+	size_t allocsz;
 } mbfl_memory_device;
 
 typedef struct _mbfl_wchar_device {
 	unsigned int *buffer;
-	int length;
-	int pos;
-	int allocsz;
+	size_t length;
+	size_t pos;
+	size_t allocsz;
 } mbfl_wchar_device;
 
-MBFLAPI extern void mbfl_memory_device_init(mbfl_memory_device *device, int initsz, int allocsz);
-MBFLAPI extern void mbfl_memory_device_realloc(mbfl_memory_device *device, int initsz, int allocsz);
+MBFLAPI extern void mbfl_memory_device_init(
+		mbfl_memory_device *device, size_t initsz, size_t allocsz);
+MBFLAPI extern void mbfl_memory_device_realloc(
+		mbfl_memory_device *device, size_t initsz, size_t allocsz);
 MBFLAPI extern void mbfl_memory_device_clear(mbfl_memory_device *device);
 MBFLAPI extern void mbfl_memory_device_reset(mbfl_memory_device *device);
-MBFLAPI extern mbfl_string * mbfl_memory_device_result(mbfl_memory_device *device, mbfl_string *result);
+MBFLAPI extern mbfl_string * mbfl_memory_device_result(
+		mbfl_memory_device *device, mbfl_string *result);
 MBFLAPI extern void mbfl_memory_device_unput(mbfl_memory_device *device);
 MBFLAPI extern int mbfl_memory_device_output(int c, void *data);
 MBFLAPI extern int mbfl_memory_device_output2(int c, void *data);
 MBFLAPI extern int mbfl_memory_device_output4(int c, void *data);
 MBFLAPI extern int mbfl_memory_device_strcat(mbfl_memory_device *device, const char *psrc);
-MBFLAPI extern int mbfl_memory_device_strncat(mbfl_memory_device *device, const char *psrc, int len);
+MBFLAPI extern int mbfl_memory_device_strncat(
+		mbfl_memory_device *device, const char *psrc, size_t len);
 MBFLAPI extern int mbfl_memory_device_devcat(mbfl_memory_device *dest, mbfl_memory_device *src);
 
 MBFLAPI extern void mbfl_wchar_device_init(mbfl_wchar_device *device);
@@ -70,4 +74,3 @@ MBFLAPI extern void mbfl_wchar_device_clear(mbfl_wchar_device *device);
 
 
 #endif /* MBFL_MEMORY_DEVICE_H */
-
